+++ ParallelGravity.h
@@ -1841,14 +1841,14 @@ public:
     void delEjected(double dDelDist, const CkCallback& cb);
     void getNearCollPartners(const CkCallback& cb);
     void getCollInfo(const CkCallback& cb);
-    void getCollInfo(int iOrder, const CkCallback& cb);
+    void getCollInfo(int64_t iOrder, const CkCallback& cb);
     void logOverlaps(const CkCallback& cb);
     void resolveCollision(Collision coll, const ColliderInfo &c1, const ColliderInfo &c2,
                           double baseStep, double timeNow, double dCentMass, const CkCallback& cb);
-    void sameHigherRung(int iord1, int rung1, int iord2, int rung2, const CkCallback& cb);
+    void sameHigherRung(int64_t iord1, int rung1, int64_t iord2, int rung2, const CkCallback& cb);
     void resolveWallCollision(Collision coll, const ColliderInfo &c1, const CkCallback& cb);
     void unKickCollStep(int iKickRung, double dDeltaBase, const CkCallback& cb);
-    void placeOnCollRung(int iOrder, int collStepRung, const CkCallback& cb);
+    void placeOnCollRung(int64_t iOrder, int collStepRung, const CkCallback& cb);
     void resetRungs(const CkCallback& cb);
     void getNeedCollStep(int collStepRung, const CkCallback& cb);
 #endif
